Preview deletion (#7)
* Revert "Make sep more explicit because can have double space elsewhere in filename"

This reverts commit dd4bd63a6d1ce12199b0c74c85e149c3e42a8631.

* Update tagmanifest for bags altered by GitHub

* Update tests with printing what is deleted

* Parse manifest with double spaces in paths

* New test and test data

* Add second argument for script mode

* Keep current functionality if mode == delete

* Print a reminder of the script mode's functionality

* Print result in preview mode

* Tests for preview mode

* Specify deleted temp are extra, not all in bag

* README add script_mode

+++ delete_new_temp.py
@@ -2,6 +2,7 @@
 
 Parameter:
     bag_path (required): path to the bag (folder that ends in "_bag")
+    script_mode (required): preview (print what will delete) or delete (actually delete)
 
 Returns:
     Prints any files that are not in the manifest but did not quality as temporary files
@@ -14,10 +15,13 @@ import re
 import sys
 
 
-def delete_temp(bag, extra_list):
+def delete_temp(bag, extra_list, mode):
     """Delete any temp files from the bag and return a list of the ones that were not deleted,
     using the same criteria as the general aip script
-    Parameter: extra_list (list) - list of paths for files that are not in the bag manifest
+    Parameters:
+        bag (string) - path to bag, needed to make full path for file
+        extra_list (list) - list of paths for files that are not in the bag manifest
+        mode (string) - preview or delete, determines if the files should just be printed or actually deleted
     Returns: not_temp (list) - list of paths for files that were not temp and therefore not deleted
     """
     not_temp = []
@@ -25,8 +29,9 @@ def delete_temp(bag, extra_list):
     for file_path in extra_list:
         file_name = file_path.split('/')[-1]
         if file_name in delete_list or file_name.endswith('.tmp') or file_name.startswith('.'):
-            print(f'Deleting {bag}/{file_path}')
-            os.remove(f'{bag}/{file_path}')
+            print(f'Delete {bag}/{file_path}')
+            if mode == 'delete':
+                os.remove(f'{bag}/{file_path}')
         else:
             not_temp.append(file_path)
     return not_temp
@@ -38,7 +43,7 @@ def find_extra_files(bag):
     Returns: extras (list) - list of the paths for every file in data but not the manifest
     """
     # List of file paths in the data folder, saved as a dataframe to compare to manifest.
-    # To match the manifest, it needs to start at data and use forward slashes.
+    # To match the manifest, the path needs to start at data and use forward slashes.
     data_paths = []
     for root, dirs, files in os.walk(os.path.join(bag, 'data')):
         for file in files:
@@ -48,14 +53,30 @@ def find_extra_files(bag):
     data_df = pd.DataFrame(data_paths, columns=['Data_Paths'])
 
     # Read the bag manifest into a dataframe.
+    # data (start of the path) has to be used as part of the separator because paths may also include a double space,
+    # and needs to be added back so future paths are correct.
     manifest_df = pd.read_csv(os.path.join(bag, 'manifest-md5.txt'), sep='  data', engine='python',
                               names=['MD5', 'Manifest_Paths'])
+    manifest_df['Manifest_Paths'] = 'data' + manifest_df['Manifest_Paths']
 
     # Compare the data path list and the bag manifest, and return those only in the data path list.
     compare_df = data_df.merge(manifest_df, left_on='Data_Paths', right_on='Manifest_Paths', how='left')
     data_only = compare_df[compare_df['Manifest_Paths'].isnull()]
     extras = data_only['Data_Paths'].tolist()
     return extras
+
+
+def reminder(mode):
+    """Prints a reminder of what the selected script mode does
+    Parameter: mode (string) - preview or delete, determines if the files should just be printed or actually deleted
+    Returns: None"""
+    if mode == 'delete':
+        print('\nRunning in script_mode "delete", which will delete extra temp files and validate the bag.')
+    elif mode == 'preview':
+        print('\nRunning in script_mode "preview", which will print files that would be deleted but changes nothing.')
+    else:
+        print(f'\nscript_mode {mode} is not "delete" or "preview".')
+        sys.exit()
 
 
 def validate_bag(bag):
@@ -75,19 +96,28 @@ def validate_bag(bag):
 if __name__ == '__main__':
     # Get bag_path from script argument.
     bag_path = sys.argv[1]
+    script_mode = sys.argv[2]
+
+    # Print reminder of script mode functionality.
+    reminder(script_mode)
 
     # Find any files that are in the bag data folder but not in the manifest (extra files).
     extra_files = find_extra_files(bag_path)
 
     # Delete any extra files that are temp files and print the path for any other files.
-    not_deleted = delete_temp(bag_path, extra_files)
+    not_deleted = delete_temp(bag_path, extra_files, script_mode)
 
-    # If all extra files were deleted, validate the bag and print the results.
+    # If the script is in delete mode and all extra files were deleted, validate the bag and print the results.
     # Otherwise, print the files that were not deleted.
-    if len(not_deleted) == 0:
-        validate_bag(bag_path)
-    else:
-        print("\nAfter deleting temp files, there are still files in the data folder that are not in the manifest:")
-        for path in not_deleted:
-            print(f'\t* {path}')
+    if script_mode == 'delete':
+        if len(not_deleted) == 0:
+            validate_bag(bag_path)
+        else:
+            print("\nAfter deleting temp files, there are still files in the data folder that are not in the manifest:")
+            for path in not_deleted:
+                print(f'\t* {path}')
+    elif script_mode == 'preview':
+        print("\nPreview of files to delete is complete.")
+        print("Files that would have been deleted are listed above.")
+        print(f"There are {len(not_deleted)} files that are not in the manifest and are not temp.")
